# api/openapi-spec/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "RESTful API for querying order information by order ID",
        "title": "wb-service API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/api/v1/orders/{orderId}": {
            "get": {
                "description": "Returns order details in JSON format.\u003cbr\u003eCheck \u003cstrong\u003eX-Cache\u003c/strong\u003e header for cache status: \u003cstrong\u003eHIT\u003c/strong\u003e (from cache) or \u003cstrong\u003eMISS\u003c/strong\u003e (from database)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Orders"
                ],
                "summary": "Get order by UID with cache status indication",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Order ID (UUID)",
                        "name": "orderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Order data",
                        "schema": {
                            "$ref": "#/definitions/github_com_Pur1st2EpicONE_WBTECH-sample-microservice_internal_models.Order"
                        },
                        "headers": {
                            "X-Cache": {
                                "type": "string",
                                "description": "Cache status: HIT or MISS"
                            }
                        }
                    },
                    "404": {
                        "description": "Order not found",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/internal_handler.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_Pur1st2EpicONE_WBTECH-sample-microservice_internal_models.Delivery": {
            "type": "object",
            "required": [
                "address",
                "city",
                "email",
                "name",
                "phone",
                "region",
                "zip"
            ],
            "properties": {
                "address": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                },
                "region": {
                    "type": "string"
                },
                "zip": {
                    "type": "string"
                }
            }
        },
        "github_com_Pur1st2EpicONE_WBTECH-sample-microservice_internal_models.Item": {
            "type": "object",
            "required": [
                "brand",
                "chrt_id",
                "name",
                "nm_id",
                "price",
                "rid",
                "status",
                "total_price",
                "track_number"
            ],
            "properties": {
                "brand": {
                    "type": "string"
                },
                "chrt_id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "nm_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "number"
                },
                "rid": {
                    "type": "string"
                },
                "sale": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "total_price": {
                    "type": "number"
                },
                "track_number": {
                    "type": "string"
                }
            }
        },
        "github_com_Pur1st2EpicONE_WBTECH-sample-microservice_internal_models.Order": {
            "type": "object",
            "required": [
                "customer_id",
                "date_created",
                "delivery",
                "delivery_service",
                "entry",
                "items",
                "locale",
                "oof_shard",
                "order_uid",
                "payment",
                "shardkey",
                "sm_id",
                "track_number"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "date_created": {
                    "type": "string"
                },
                "delivery": {
                    "$ref": "#/definitions/github_com_Pur1st2EpicONE_WBTECH-sample-microservice_internal_models.Delivery"
                },
                "delivery_service": {
                    "type": "string"
                },
                "entry": {
                    "type": "string"
                },
                "internal_signature": {
                    "type": "string"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_Pur1st2EpicONE_WBTECH-sample-microservice_internal_models.Item"
                    }
                },
                "locale": {
                    "type": "string"
                },
                "oof_shard": {
                    "type": "string"
                },
                "order_uid": {
                    "type": "string"
                },
                "payment": {
                    "$ref": "#/definitions/github_com_Pur1st2EpicONE_WBTECH-sample-microservice_internal_models.Payment"
                },
                "shardkey": {
                    "type": "string"
                },
                "sm_id": {
                    "type": "integer"
                },
                "track_number": {
                    "type": "string"
                }
            }
        },
        "github_com_Pur1st2EpicONE_WBTECH-sample-microservice_internal_models.Payment": {
            "type": "object",
            "required": [
                "amount",
                "currency",
                "delivery_cost",
                "goods_total",
                "payment_dt",
                "provider",
                "transaction"
            ],
            "properties": {
                "amount": {
                    "type": "number"
                },
                "bank": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "custom_fee": {
                    "type": "number"
                },
                "delivery_cost": {
                    "type": "number"
                },
                "goods_total": {
                    "type": "number"
                },
                "payment_dt": {
                    "type": "integer"
                },
                "provider": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                },
                "transaction": {
                    "type": "string"
                }
            }
        },
        "internal_handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        }
    }
}